+++ TOMB4/specific/function_stubs.h
@@ -6,15 +6,15 @@ void SeedRandomControl(long seed);
 long GetRandomDraw();
 void SeedRandomDraw(long seed);
 void init_game_malloc();
-void* game_malloc(long size);
+void* game_malloc(size_t size);
 
 #ifndef DEBUG
 #define SYSTEM_MALLOC(size) malloc(size)
 #define SYSTEM_REALLOC(ptr, size) realloc(ptr, size)
 #define SYSTEM_FREE(ptr) free(ptr)
 #else
-void* system_malloc(long size, const char* filename, int line_number);
-void* system_realloc(void* ptr, long size, const char* filename, int line_number);
+void* system_malloc(size_t size, const char* filename, int line_number);
+void* system_realloc(void* ptr, size_t size, const char* filename, int line_number);
 void system_free(void* ptr);
 #define SYSTEM_MALLOC(size) system_malloc(size, __FILE__, __LINE__)
 #define SYSTEM_REALLOC(ptr, size) system_realloc(ptr, size, __FILE__, __LINE__)
